Workaround SpinBox's focus handling flaws
There is a flaw in QQC2 SpinBox with how focus is handled.
Unless you make the TextInput/TextField unfocusable (not what you want),
you can't detect whether or not a SpinBox has activeFocus or
visualFocus.
This is because the TextInput takes the focus.
In order to detect if a SpinBox is focused, you must trust that the
contentItem will be the part that takes the focus.