Add IconLabelLayout
IconLabelLayout is based on QtQuick.Controls.impl IconLabel. Instead of
using the relevant icon and label C++ classes directly to create the
icon and label, this uses QML components and creates them in C++ as
QQuickItems. I have to do it this way because the C++ classes for
Kirigami Icon, Qt Quick Image, Qt Quick Text and QQC2 Label are all
private. Remarkably, it seems work just as well as Qt's IconLabel.

Also adjust other components to be compatible with it.