Port to the new C++ units plugin in Kirigami
This adds a C++ Kirigami Units instance. It also moves the Units.qml
file that from the Kirigami directory to the impl directory, adjusts it
to be based on Kirigami.Units and then adjusts all files to make use of
the Units in impl. This is done to avoid having to add all the custom
properties to the C++ plugin.