Fix padding for items without indicators in menus with
When a MenuItem does not have an indicator, in a Menu with other
items which do have indicators, the padding is incorrect, so the
text is not aligned properly. Use the same value as the indicators
for padding, even when the control doesn't have an indicator, if
the menu has other items with indicators, to keep the text aligned.