Bundle color definitions on Android
On Android we want to be able to switch between a light and dark color
scheme. To do that we need the .color files.

Usually they are shipped with the breeze repo, but that is not used at
all on Android.

By installing them the files become available as assets in the
application and (a patched) KColorSchemeModel can find them.

Since these will be the only two schemes on Android just call them Light and Dark

This way the files are duplicated in our code, but I can't think of a
better alternative.