Avoid needlessly reading font settings
Currently we call syncFonts every time a platformtheme is created and updated, which is way too often.

It is all global data, i.e. not specific to the concrete platformtheme instance, so move it to the style singleton.

Also only read the settings for smallFont when fonts were actually changed.