MenuItem: Fix TypeError: Cannot read property '*Padding' of null
The problem was setting `checkIndicator.control` after the component had
been loaded, so its properties were trying to be read when it was still
null.

It seems we cannot simply write `control: control` inside of a Component
due to scoping issues. It results in a binding loop error. Instead we
define the alias `controlRoot` and use that.