More thoroughly fix the QQuickIcon name clash
The copy is a distinct type and needs distinct name for things to work,
otherwise QML might give us one with the memory layout of the other. If
we are lucky that happens to work, but on Android/ARM32 this reliably
triggers a SIGBUG/BUS_ADRALN alignment violation.

To solve this, put the copy into our own namespace.