Fix vertical slider groove alignment
For some reason 'x' gets reset to 0 if it's defined after implicitWidth
and/or width.  This might be a bug in qt.
This adapted implementation corresponds to the implementation in Fusion
and Material style, where x and y are also set before implicitWidth
and width, and implicitHeight and height, respectively.